Homework_4 fixed
