+++ src/main/java/hw4/common/utils/Utils.java
@@ -1,5 +1,8 @@
 package hw4.common.utils;
 
+import com.codeborne.selenide.CollectionCondition;
+import com.codeborne.selenide.impl.WebElementsCollection;
+import org.checkerframework.checker.nullness.compatqual.NullableDecl;
 import org.openqa.selenium.WebElement;
 
 import java.util.ArrayList;
@@ -31,5 +34,29 @@ public class Utils {
             converted.append(ch);
         }
         return converted.toString();
+    }
+
+    public static CollectionCondition searchedValueInEachRecord(String searchingWord) {
+        return new CollectionCondition() {
+            @Override
+            public void fail(WebElementsCollection webElementsCollection, List<WebElement> list, Exception e, long l) {
+                e.printStackTrace();
+            }
+
+            @Override
+            public boolean applyNull() {
+                return false;
+            }
+
+            @Override
+            public boolean apply(@NullableDecl List<WebElement> webElements) {
+                for (WebElement element : webElements) {
+                    if (!element.getText().contains(searchingWord)) {
+                        return false;
+                    }
+                }
+                return true;
+            }
+        };
     }
 }
